requirements.txt
PyPDF2
langchain
langchain_community 
langchain_openai
langchain_core 
streamlit
python-dotenv
langserve
flask 
chromadb
sentence-transformers
groq
langchain-groq 
langchain_google_genai 
streamlit_chat
faiss-cpu
streamlit-extras
unstructured
langchain-huggingface
streamlit_pdf_viewer
streamlit-authenticator